### 2025-04-29 #8137 @editorialbot
[PRE REVIEW]: Automated Scaling of PersistentVolumeClaims in Kubernetes: Design and Implementation of the VolumeScaler Controller
**Submitting author:** @zghanem0 (<a href="http://orcid.org/0009-0005-0012-4470">Ahmed Ghanem</a>)
**Repository:** https://github.com/aws-samples/sample-volumeScaler
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.6
