+++ paper/paper.md
@@ -25,6 +25,18 @@ The management of storage resources in Kubernetes clusters is a critical aspect 
 
 `VolumeScaler` was designed to be used by both cloud platform administrators and application developers managing storage resources in Kubernetes environments. It has been implemented using Kubernetes Custom Resource Definitions (CRDs) and integrates with Container Storage Interface (CSI) drivers [@csi_spec_2022] (`storage.k8s.io` and `csi.storage.k8s.io`). The combination of automation, flexibility, and support for Kubernetes storage functionality in `VolumeScaler` will enable efficient management of storage resources in dynamic cloud environments. The source code for `VolumeScaler` is available on GitHub.
 
+
+# Statement of need
+
+While Kubernetes provides comprehensive solutions for scaling compute resources through features like Horizontal Pod Autoscaling (HPA) [@HPA], the platform lacks native mechanisms for automatically scaling storage resources. This gap in functionality becomes particularly challenging in dynamic cloud environments where storage requirements can fluctuate significantly. Manual intervention for storage scaling operations introduces several challenges:
+
+1. Operational overhead: Administrators must manually monitor storage usage and perform scaling operations
+2. Resource inefficiency: Static storage allocation often leads to either over-provisioning (increasing costs) or under-provisioning (risking performance degradation)
+3. Response latency: Manual scaling operations cannot respond quickly to sudden changes in storage requirements
+
+`VolumeScaler` addresses these challenges by providing automated, policy-driven scaling of PersistentVolumeClaims. The controller continuously monitors storage usage patterns and automatically adjusts PVC sizes based on configurable thresholds and policies. This automation reduces operational overhead while ensuring optimal resource utilization. The implementation follows Kubernetes best practices [@k8s_controllers_2021] and integrates seamlessly with existing storage solutions through the Container Storage Interface [@csi_spec_2022].
+
+
 # Acknowledgements
 
 I would like to acknowledge the support and guidance from my supervisors and colleagues at Munster Technological University Cork. Special thanks to the Kubernetes community for their continuous development of the platform and its ecosystem.
