### 2025-04-29 #8137 @danielskatz
👋 @zghanem0 - Thanks for your submission.

1. Please add a References header at the end of the .md file, as seen in our [example paper](https://joss.readthedocs.io/en/latest/example_paper.html).
2. Please add a statement of need section in the paper, as in the example paper. Currently, this paper does not mention research, while we say "The software must have an obvious research application" in https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements
3. Please add the country to your affiliation
4. In your bib file, you will need to add an extra set of curly braces around the authors of the two references that are in the paper so that they are not interpreted as names.

Please make changes to your .bib and .md files, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
