+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{KubernetesResourceManagement,
-    author    = {Kubernetes Documentation},
+    author    = {{Kubernetes Documentation}},
     title     = {Resource Management},
     journal   = {Kubernetes Official Documentation},
     year      = {2024},
@@ -7,7 +7,7 @@
 }
 
 @article{k8s_storage_2023,
-    author = {Kubernetes Storage SIG},
+    author = {{Kubernetes Storage SIG}},
     title = {Kubernetes Storage Architecture and Future Directions},
     journal = {USENIX ;login:},
     year = {2023},
@@ -17,7 +17,7 @@
 }
 
 @inproceedings{csi_spec_2022,
-    author = {Container Storage Interface Authors},
+    author = {{Container Storage Interface Authors}},
     title = {Container Storage Interface Specification},
     booktitle = {Cloud Native Computing Foundation},
     year = {2022},
