+++ paper/paper.bib
@@ -29,4 +29,12 @@
     title = {Building Custom Controllers for Kubernetes},
     booktitle = {Cloud Native Computing Conference},
     year = {2021}
+}
+
+@article{HPA,
+    author = {{Kubernetes Documentation}},
+    title = {Horizontal Pod Autoscaling},
+    journal = {Kubernetes Official Documentation},
+    year = {2024},
+    url = {https://kubernetes.io/docs/tasks/run-application/horizontal-pod-autoscale/}
 } 
