### 2025-05-07 #8137 @danielskatz
Thanks!!  For the future, you can just accept directly by using the command above, or even just `@editorialbot assign me as editor`
