+++ volumescaler.yaml
@@ -159,12 +159,12 @@ spec:
             allowPrivilegeEscalation: true
           volumeMounts:
             - name: kubelet
-              mountPath: /var/lib/kubelet
+              mountPath: /var/lib/kubelet/pods
               mountPropagation: Bidirectional
       volumes:
         - name: kubelet
           hostPath:
-            path: /var/lib/kubelet
+            path: /var/lib/kubelet/pods
             type: Directory
       dnsPolicy: ClusterFirst
       nodeSelector:
