### 2025-05-11 a4ed0d6
Update cr.yaml configuration
