+++ cr.yaml
@@ -1,4 +1,5 @@
 owner: aws-samples
 git-repo: sample-volumeScaler
 charts-repo: https://aws-samples.github.io/sample-volumeScaler
-packages-with-index: true 
+packages-with-index: true
+charts-dir: charts/volumescaler 
