+++ .github/cr.yaml
@@ -1,3 +1,1 @@
-
-
 release-name-template: "helm-chart-{{ .Name }}-{{ .Version }}"
