+++ test-pod-data-generator.yaml
@@ -3,7 +3,6 @@ apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
   name: test-ebs-claim
-  
 spec:
   accessModes:
     - ReadWriteOnce
@@ -16,7 +15,6 @@ apiVersion: autoscaling.storage.k8s.io/v1alpha1
 kind: VolumeScaler
 metadata:
   name: test-ebs-claim
-  
 spec:
   pvcName: test-ebs-claim
   threshold: "70%"
