### 2025-05-20 4334c69
adding how it works image
