+++ paper/paper.md
@@ -10,10 +10,13 @@ authors:
   - name: Ahmed Ghanem
     orcid: 0009-0005-0012-4470
     affiliation: 1
+  - name: Ruairi O'Reilly
+    orcid: 0000-0001-7990-3461
+    affiliation: 2
 affiliations:
  - name: Department of Computer Science, Munster Technological University Cork, Ireland
    index: 1
-date: 23 April 2024
+date: 23 April 2025
 bibliography: paper.bib
 ---
 
