### 2025-08-01 #8137 @danielskatz
👋 @bkrayfield - can you keep working on finding reviewers, to get this started? Feel free to use the JOSS slack to get advice or suggestions.
