### 2025-08-10 #8137 @akshaymittal143
> 👋 [@akashkthkr](https://github.com/akashkthkr) or [@akshaymittal143](https://github.com/akshaymittal143) can you review this submission for JOSS?

Yes, I can. This will be my first, so I need some reviewer guidelines, but I should be able to help!
