### 2025-08-12 #8137 @bkrayfield
@akshaymittal143 Thank you for stepping up! I’ll be here to support you along the way. In the meantime, you may want to start by reviewing the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). We are still in the pre-review stage and looking for a second reviewer. Once we have one, the review will begin and you can generate your checklist.
