### 2025-08-13 ffa97e9
Address JOSS reviewer feedback: Add State of Field section, fix citations, add GoDoc comments, and emphasize CSI driver compatibility
