+++ paper/paper.bib
@@ -38,3 +38,42 @@
     year = {2024},
     url = {https://kubernetes.io/docs/tasks/run-application/horizontal-pod-autoscale/}
 } 
+
+@techreport{velero_2022,
+    author = {VMware},
+    title = {Velero Technical Architecture},
+    institution = {VMware, Inc.},
+    year = {2022},
+    type = {Technical Report}
+}
+
+@misc{stork_docs,
+    author = {Portworx},
+    title = {STORK: Storage Operator Runtime for Kubernetes},
+    year = {2023},
+    howpublished = {\url{https://github.com/libopenstorage/stork}},
+    note = {Accessed: 2023-12-01}
+}
+
+@inproceedings{konev2022,
+    author = {Konev, Igor and Nikiforov, Igor and Ustinov, Sergey},
+    booktitle = {2022 31st Conference of Open Innovations Association (FRUCT)}, 
+    title = {Algorithm for Containers' Persistent Volumes Auto-scaling in Kubernetes}, 
+    year = {2022},
+    volume = {},
+    number = {},
+    pages = {89-95},
+    keywords = {Technological innovation;Costs;Automation;Clustering algorithms;Manuals;Containers;Computer crashes},
+    doi = {10.23919/FRUCT54823.2022.9770916}
+}
+
+@inproceedings{In-Memory_Storage,
+    author = {Shrestha, Bibek R. and Mitra, Saptashwa and Pallickara, Sangmi L.},
+    year = {2019},
+    title = {STRETCH: In-Memory Storage with Autoscaling for Cluster Computing},
+    booktitle = {IEEE International Conference on Cluster Computing},
+    publisher = {IEEE},
+    pages = {224-228},
+    abstract = {IoT enabled devices and observational instruments continuously generate voluminous data. A large portion of these datasets are delivered with the associated geospatial locations. The increased volumes of geospatial data, alongside the emerging geospatial services, pose computational challenges for large-scale geospatial analytics. We have designed and implemented STRETCH, an in-memory distributed geospatial storage that preserves spatial proximity and enables proactive autoscaling for frequently accessed data. STRETCH stores data with a delayed data dispersion scheme that incrementally adds data nodes to the storage system. STRETCH is built on Apache Spark and Ignite and interacts with them at runtime.},
+    keywords = {Geospatial Analysis; In memory Cluster Computing; Scalable Analytics}
+} 
