+++ cmd/main.go
@@ -100,6 +100,16 @@ type ControllerConfig struct {
 }
 
 // NewDefaultConfig returns a default controller configuration
+// NewDefaultConfig returns a default controller configuration with predefined values
+// suitable for most production environments.
+//
+// Default values:
+//   - KubeletPodsPath: "/var/lib/kubelet/pods" (standard kubelet path)
+//   - PollInterval: 60 seconds (reasonable balance between responsiveness and resource usage)
+//   - MaxRetries: 3 (sufficient for transient failures)
+//   - Timeout: 30 seconds (adequate for most Kubernetes API operations)
+//
+// Returns a configured ControllerConfig ready for use.
 func NewDefaultConfig() *ControllerConfig {
 	return &ControllerConfig{
 		KubeletPodsPath: defaultKubeletPodsPath,
@@ -116,6 +126,13 @@ func NewDefaultConfig() *ControllerConfig {
 // -------------------------------------
 // 2) convertToGi: converts "5Gi" / "512Mi" / "1Ti" into float64 Gi
 // -------------------------------------
+// convertToGi parses storage size strings and converts them to GiB units.
+// Supports common storage units: Mi (mebibytes), Gi (gibibytes), Ti (tebibytes).
+//
+// Parameters:
+//   - sizeStr: Storage size string (e.g., "5Gi", "512Mi", "1Ti")
+//
+// Returns the size in GiB as a float64, or an error if parsing fails.
 func convertToGi(sizeStr string) (float64, error) {
 	if sizeStr == "" {
 		return 0, fmt.Errorf("empty size string")
@@ -156,6 +173,14 @@ func convertToGi(sizeStr string) (float64, error) {
 // -------------------------------------
 // 3) inClusterOrKubeconfig: tries in-cluster config first, fallback to KUBECONFIG
 // -------------------------------------
+// inClusterOrKubeconfig attempts to create a Kubernetes client configuration
+// by first trying to use in-cluster configuration (when running inside a pod),
+// then falling back to the KUBECONFIG environment variable if available.
+//
+// This function enables the controller to work both when deployed as a DaemonSet
+// in a cluster and when running locally for development/testing.
+//
+// Returns a valid Kubernetes client configuration, or an error if both methods fail.
 func inClusterOrKubeconfig() (*rest.Config, error) {
 	kubeconfig := os.Getenv("KUBECONFIG")
 	if kubeconfig == "" {
@@ -176,6 +201,19 @@ func inClusterOrKubeconfig() (*rest.Config, error) {
 // ------------------------------------------------------
 // 4) getPVCUIDsFromLocalMounts: returns slice of PVC UIDs + map PVC UID -> mountPath
 // ------------------------------------------------------
+// getPVCUIDsFromLocalMounts discovers PVCs mounted on the current node by scanning
+// the kubelet pods directory for CSI volume mounts. This function is essential for
+// the DaemonSet architecture, as it allows each node to identify only the PVCs
+// that are relevant to that specific node.
+//
+// Parameters:
+//   - kubeletPodsPath: Path to the kubelet pods directory (typically "/var/lib/kubelet/pods")
+//   - clientset: Kubernetes client for additional validation if needed
+//
+// Returns:
+//   - Slice of PVC UIDs found on the current node
+//   - Map from PVC UID to its mount path
+//   - Error if the discovery process fails
 func getPVCUIDsFromLocalMounts(kubeletPodsPath string, clientset kubernetes.Interface) ([]types.UID, map[types.UID]string, error) {
 	pvcUIDSet := make(map[types.UID]struct{})
 	pvcUIDToMount := make(map[types.UID]string)
@@ -228,6 +266,20 @@ func getPVCUIDsFromLocalMounts(kubeletPodsPath string, clientset kubernetes.Inte
 // ---------------------------------------------
[…]
