### 2025-08-14 #8137 @bkrayfield
Hello @akshaymittal143,

Thank you for your quick and thoughtful comments. JOSS uses a checklist-style review process, and while you may have already addressed many of these items, we are still in the process of securing a second reviewer. I appreciate your patience as we work to bring another reviewer on board to contribute alongside you.
