### 2025-09-22 #8846 @danielskatz
@akshaymittal143 - JOSS does not require anything like what you say, and doesn't supply information to any automated systems currently, but if you need a letter confirming your review activity, we can provide such a letter.
