### 2025-10-16 66ff0ba
feat: add CONTRIBUTING.md and GitHub Actions CI/CD pipeline
- Add comprehensive CONTRIBUTING.md with development guidelines
- Add GitHub Actions workflow for automated testing, linting, and building
- Include multi-arch Docker builds and security scanning
- Address JOSS review feedback for contribution guidelines and CI/CD
