+++ CONTRIBUTING.md
@@ -1,0 +1,225 @@
+# Contributing to VolumeScaler
+
+Thank you for your interest in contributing to VolumeScaler! This document provides guidelines and instructions for contributing to the project.
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [Getting Started](#getting-started)
+- [Development Setup](#development-setup)
+- [Making Changes](#making-changes)
+- [Testing](#testing)
+- [Submitting Changes](#submitting-changes)
+- [Release Process](#release-process)
+
+## Code of Conduct
+
+This project adheres to a code of conduct that we expect all contributors to follow. Please be respectful and professional in all interactions.
+
+## Getting Started
+
+### Prerequisites
+
+- Go 1.23 or later
+- Docker and Docker Buildx
+- kubectl configured with access to a Kubernetes cluster
+- A Kubernetes cluster with CSI drivers that support volume expansion (see [CSI Drivers](https://kubernetes-csi.github.io/docs/drivers.html))
+
+### Fork and Clone
+
+1. Fork the repository on GitHub
+2. Clone your fork locally:
+   ```bash
+   git clone https://github.com/YOUR_USERNAME/sample-volumeScaler.git
+   cd sample-volumeScaler
+   git remote add upstream https://github.com/aws-samples/sample-volumeScaler.git
+   ```
+
+## Development Setup
+
+### Local Development
+
+1. **Install dependencies:**
+   ```bash
+   go mod tidy
+   ```
+
+2. **Build the project:**
+   ```bash
+   make build
+   ```
+
+3. **Run tests:**
+   ```bash
+   make test
+   ```
+
+4. **Generate coverage report:**
+   ```bash
+   make coverage
+   ```
+
+### Testing Environment
+
+For comprehensive testing, we recommend using a production-grade Kubernetes cluster with proper CSI drivers:
+
+- **AWS EKS** with EBS CSI driver
+- **Azure AKS** with Azure Disk CSI driver  
+- **Google GKE** with Persistent Disk CSI driver
+- **Production clusters** with any [supported CSI drivers](https://kubernetes-csi.github.io/docs/drivers.html)
+
+**Note:** VolumeScaler is designed for production environments and requires CSI drivers that support online volume expansion. Lightweight distributions like k3s may not provide the full CSI functionality needed for proper testing.
+
+## Making Changes
+
+### Branch Strategy
+
+1. Create a feature branch from `main`:
+   ```bash
+   git checkout -b feature/your-feature-name
+   ```
+
+2. Make your changes following the coding standards below
+
+3. Commit your changes with clear, descriptive messages:
+   ```bash
+   git commit -m "feat: add new scaling algorithm for predictable workloads"
+   ```
+
+### Coding Standards
+
+- Follow Go best practices and conventions
+- Write clear, self-documenting code
+- Add unit tests for new functionality
+- Update documentation as needed
+- Use meaningful variable and function names
+
+### Commit Message Format
+
+We follow conventional commit format:
+
+```
+type(scope): description
+
+[optional body]
+
+[optional footer]
+```
+
+Types: `feat`, `fix`, `docs`, `style`, `refactor`, `test`, `chore`
+
+## Testing
+
+### Unit Tests
+
+Run the full test suite:
+```bash
+make test
+```
+
+### Integration Testing
+
+For integration testing, deploy VolumeScaler to a cluster with proper CSI drivers:
+
+1. **Deploy VolumeScaler:**
+   ```bash
+   kubectl apply -f volumescaler.yaml
+   ```
+
+2. **Deploy test workload:**
+   ```bash
+   kubectl apply -f test-pod-data-generator.yaml
[…]
