### 2025-10-16 bab082d
fix: resolve GitHub Actions workflow failures
- Fix golangci-lint installation by using single action
- Simplify YAML validation to avoid conflicts with existing workflows
- Remove integration test job that was causing issues
- Focus workflow on main branch only to avoid conflicts
