### 2025-10-16 1e9825b
fix: disable Docker build and release jobs to focus on core CI
- Disable Docker build job since ECR credentials aren't configured
- Disable release job for now to focus on essential CI checks
- Keep linting, testing, validation, and security scanning active
- Workflow now focuses on code quality and validation
