### 2025-10-16 ec67af9
feat: replace complex CI with simple, focused workflow
- Replace problematic complex CI workflow with simple-ci.yml
- Focus on essential checks: linting, testing, validation, building
- Remove Docker builds and security scans that require external setup
- Streamlined workflow that actually works and provides value
