### 2025-12-11 4dc5996
ci: scope down permissions for release-chart.yaml
