+++ .github/workflows/release-chart.yaml
@@ -9,6 +9,9 @@ on:
       - '.github/workflows/release-chart.yaml'
       - 'cr.yaml'
 
+permissions:
+  contents: write
+
 jobs:
   release:
     runs-on: ubuntu-latest
