### 2025-12-16 cb0c60b
fix: scope GitHub Actions token permissions to least privilege
- Add explicit 'contents: read' permission to publish.yml workflow
- Add 'contents: write' and 'pages: write' permissions to release-chart.yaml workflow
- Addresses security finding for GitHub token permission scoping

Resolves penetration testing security ticket for overly permissive GitHub Actions token permissions.
