### 2025-12-16 cd23694
fix: remove conflicting write-all permissions from workflows
- Remove 'permissions: write-all' from release.yaml and helm-chart-release.yaml
- Keep only the explicitly scoped permissions at workflow level
- Ensures consistent least-privilege permission model across all workflows
