+++ .github/workflows/helm-chart-release.yaml
@@ -28,7 +28,6 @@ permissions:
 jobs:
   release:
     runs-on: ubuntu-latest
-    permissions: write-all
     steps:
       - name: Checkout
         uses: actions/checkout@v4
