+++ .github/workflows/release.yaml
@@ -26,7 +26,6 @@ jobs:
   build:
     name: Release
     runs-on: ubuntu-latest
-    permissions: write-all
     steps:
       - name: Create Release
         id: create-release
