### 2025-12-16 607ffa2
fix: add skip_existing flag to chart-releaser workflows
- Prevents failures when trying to create releases that already exist
- Allows workflows to complete successfully when chart version hasn't changed
- Fixes the 422 Validation Failed error for existing release tags
