+++ .github/workflows/release-chart.yaml
@@ -41,4 +41,6 @@ jobs:
         uses: helm/chart-releaser-action@v1.6.0
         env:
           CR_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
-          CR_CHARTS_DIR: "charts/volumescaler" 
+          CR_CHARTS_DIR: "charts/volumescaler"
+        with:
+          skip_existing: true 
