### 2026-01-30 2a116f1
Add Results and Analysis section with plots to paper.md
