### 2026-03-10 #8846 @bkrayfield
https://github.com/openjournals/joss-reviews/issues/8846#issuecomment-3483689876
