### 2026-03-10 #8846 @danielskatz
Ok, thanks, though the author isn't in charge of changing reviewers, which is the editor's job. Hopefully @akashkthkr will reply and confirm this one way or the other, but I'll look for another reviewer in a few days if he agrees with this or I don't hear anything.
