### 2026-03-17 #8846 @danielskatz
👋 @shrinandj & @asmacdo & @SergeyYakubov & @Ingvord & @summitsingh & @kedark3, would one of you be willing to review this submission for JOSS? We have one complete review, but lost our second reviewer, so need another. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
