### 2026-03-17 #8846 @danielskatz
Thanks @kedark3 - I'll add you now.  You can then use `@editorialbot generate my checklist` as a command to generate your review checklist, but it will be the current one, while this review should be done under our prior guidelines.  I'll edit the checklist for you to bring it back to the old one. Please see https://github.com/openjournals/joss-reviews/issues/8846#issuecomment-3719121370 for discussion about this scope change, and the fact that this review is under the old scope.
