### 2026-03-17 #8846 @danielskatz
@kedark3 - I've updated your checklist to the appropriate items for a pre-2026 submission. If you checked anything off while I was editing, you'll need to recheck that item in the current list. I apologize if this happened.
