### 2026-03-17 #8846 @kedark3
@danielskatz I am not sure how to check these two off:

> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
>
> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

Cause there are plots in the paper, but do not see any data that backs those plots. And also results are produced over a period of 30days.

The method is described on how the results were produced, so I take that as a way to reproduce the results, so I think we can check off the reproducibility.  Agree?
