### 2026-03-17 #8846 @danielskatz
👋 @zghanem0 - What's your opinion on the questions from @kedark3 above?
