### 2026-03-17 #8846 @kedark3
I also tried installing the operator and testing it on a `KIND` cluster but failed with error:
```
  Warning  FailedScheduling  49s   default-scheduler  0/1 nodes are available: pod has unbound immediate PersistentVolumeClaims. not found
```

Most people don't have a full fledged Kube cluster at their disposal, can you include instructions to run this on `KIND` or `Minikube` cluster? @zghanem0

I was able to just create a fake `gp2` class as below:
```yaml
kubectl apply -f - <<EOF
apiVersion: storage.k8s.io/v1
kind: StorageClass
metadata:
  name: gp2
provisioner: k8s.io/minikube-hostpath
EOF
```

Then reapplying test manifest worked. Pod came up . But it did not autoscale the PVC. I request you to provide steps to test it in Minikube or KIND cluster.
