### 2026-03-23 #8846 @zghanem0
Hi @kedark3 , thanks for the feedback!

The README now includes full step-by-step deployment instructions for Minikube and k3s. For Minikube, the key is enabling the csi-hostpath-driver addon and using csi-hostpath-sc as the StorageClass (with allowVolumeExpansion: true patched in). The default standard class and a fake gp2 class using k8s.io/minikube-hostpath don't report volume statistics to the kubelet, which is why the controller never received usage data in your test.

Please make sure you're testing with the latest version (v0.2.0) — pull the latest from main or use the image public.ecr.aws/ghanem/volumescaler:v0.2.0.

See the updated README: https://github.com/aws-samples/sample-volumeScaler?tab=readme-ov-file#deploying-on-minikube
