### 2026-03-27 #8846 @kedark3
I still need answers for: https://github.com/openjournals/joss-reviews/issues/8846#issuecomment-4078522762  . Once I have the answers, we will complete the review. Everything else checks out.
I was able to test the installation and functionality with Minikube, so that works well, I can confirm.
