### 2026-03-27 #8846 @zghanem0
Hello @kedark3,  Thanks for raising this. The datasets backing the plots have now been added to the repository under paper/data/. These contain the daily metrics collected during the 30-day evaluation on Amazon EKS with EBS-backed PVCs (gp2 StorageClass), covering cost, provisioning, and reliability for both the manual and automated scaling scenarios.

An analysis script (analysis_code.py) is also included, running it against the provided CSVs reproduces all figures in the paper.
