### 2026-03-31 #8846 @danielskatz
👋 @zghanem0 - now that we have 2 complete reviews, we can continue with the last steps of the process.  I was going to proofread the paper, but I see that there are some problems.

1. There's a DOI that should be added to a reference, as indicated above
2. The PDF doesn't compile.  The error above appears to be `Problem with affiliations for Ruairi O'Reilly, perhaps the affiliations index need quoting?`

Can you work on these issues? You can use the commands I used to check the references and generating the PDF again when you are ready.
