### 2026-03-31 #8846 @danielskatz
@zghanem0 - I've proofread the paper now, and suggest the changes in https://github.com/aws-samples/sample-volumeScaler/pull/9 - please merge this or let me know what you disagree with.  Also, the acknowledgments starts with "I", but there are two authors, so it's unclear who "I" is. Please either change this to "we" or identify which author is speaking.
