### 2026-03-31 #8846 @zghanem0
@danielskatz Done: https://zenodo.org/records/19358444

I checked the following:
- Tagged release: v0.2.0
- Zenodo archive DOI:  zenodo.19358444
- Metadata verified: title matches the paper ("VolumeScaler: A Kubernetes Controller for Automated PVC Scaling"), authors are Ahmed Ghanem (ORCID: 0009-0005-0012-4470) and Ruairi O'Reilly (ORCID: 0000-0001-7990-3461).
