### 2026-03-31 #8846 @danielskatz
Congratulations to @zghanem0 (Ahmed Ghanem) and co-author on your JOSS publication!!

And thanks to @akshaymittal143 & @kedark3 for reviewing!

JOSS completely depends on volunteers - we couldn't do this without you
